# Conflicts:
#	FileApi/Api/Api.csproj
#	FileApi/Api/Controllers/File/Dto/Request/CreateFileRequest.cs
#	FileApi/Api/Controllers/File/Dto/Request/DeleteFileRequest.cs
#	FileApi/Api/Controllers/File/Dto/Request/GetFileRequest.cs
#	FileApi/Api/Controllers/File/Dto/Request/UpdateFileRequest.cs
#	FileApi/Api/Controllers/File/Dto/Response/CreateFileResponse.cs
#	FileApi/Api/Controllers/File/Dto/Response/GetFileListResponse.cs
#	FileApi/Api/Controllers/File/Dto/Response/GetFileResponse.cs
#	FileApi/Api/Controllers/File/FileController.cs
#	FileApi/Api/Controllers/File/Profiles/FileMappingProfile.cs
#	FileApi/Api/Program.cs
#	FileApi/Api/Startup.cs
#	FileApi/Dal/ForStartup.cs
#	FileApi/Dal/Helpers/FileHelper.cs
#	FileApi/Dal/Migrations/Date202402272241_AddTable_FileDal.cs
#	FileApi/Dal/Migrations/InitialMigration.cs
#	FileApi/Dal/Models/FileDal.cs
#	FileApi/Dal/Repositories/FileRepository/FileRepository.cs
#	FileApi/Dal/Repositories/FileRepository/Interfaces/IFileRepository.cs
#	FileApi/FileApi.sln
#	FileApi/Logic/ForStartup.cs
#	FileApi/Logic/Managers/File/Managers/FileManager.cs
#	FileApi/Logic/Managers/File/Managers/Interfaces/IFileManager.cs
#	FileApi/Logic/Models/FileModel.cs
#	FileApi/Logic/Records/File/Params/CreateFileParam.cs
#	FileApi/Logic/Records/File/Params/UpdateFileParam.cs
#	FileApi/Logic/Records/File/Results/CreateFileResult.cs
#	FileApi/Logic/Records/File/Results/GetFileListResult.cs
#	FileApi/Logic/Records/File/Results/GetFileResult.cs
#	FileApi/Logic/Services/Interfaces/IFileStorageService.cs
#	FileApi/Logic/Services/Storages/YandexFileStorageService.cs
#	Libs/Core/Core.csproj
#	Libs/Core/Dal/Base/BaseEntityDal.cs
#	Libs/Core/Dapper/Connection/DapperContext.cs
#	ModelPredictionApi/Api/Api.csproj
#	ModelPredictionApi/Api/Startup.cs
#	ModelPredictionApi/Logic/Logic.csproj
#	ModelPredictionApi/ModelPredictionApi.sln

+++ FileApi/Logic/Managers/File/Managers/FileManager.cs
@@ -1,4 +1,6 @@
 ﻿using AutoMapper;
+using Core.Exceptions;
+using Dal.Enums;
 using Dal.Helpers;
 using Dal.Models;
 using Dal.Repositories.FileRepository.Interfaces;
@@ -8,16 +10,16 @@ using Logic.Records.File.Params;
 using Logic.Records.File.Results;
 using Logic.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
+using FileNotFoundException = Core.Exceptions.FileNotFoundException;
 
 namespace Logic.Managers.File.Managers;
 
 /// <inheritdoc cref="IFileManager"/>
 internal class FileManager : IFileManager
 {
-    private IFileRepository _fileRepository;
-    private IFileStorageService _fileStorageService; 
-    private IMapper _mapper;
+    private readonly IFileRepository _fileRepository;
+    private readonly IFileStorageService _fileStorageService; 
+    private readonly IMapper _mapper;
 
     public FileManager(IFileRepository fileRepository, IFileStorageService fileStorageService, IMapper mapper)
     {
@@ -27,10 +29,10 @@ internal class FileManager : IFileManager
     }
     
     /// <inheritdoc cref="IFileManager.UploadFileAsync"/>
-    public async Task<CreateFileResult> UploadFileAsync(CreateFileParam createFileParam, IFormFile file)
+    public async Task<CreateFileResult> UploadFileAsync(IFormFile file)
     {
-        var fileName = FileHelper.GetFileNameFromFileNameWithExtension(file.FileName);
-        var fileExtension = FileHelper.GetExtensionFile(file.FileName);
+        var fileName = Path.GetFileNameWithoutExtension(file.FileName);
+        var fileExtension = Path.GetExtension(file.FileName);
         
         var fileDal = new FileDal
         {
@@ -39,7 +41,7 @@ internal class FileManager : IFileManager
             FileExtension = fileExtension,
             CreatedTime = DateTime.Now,
             FileSize = file.Length,
-            AccessModifier = createFileParam.AccessModifier
+            AccessModifier = AccessModifier.Private
         };
 
         var fileId = await _fileRepository.SaveFileAsync(fileDal);
@@ -56,11 +58,22 @@ internal class FileManager : IFileManager
     {
         var fileDal = await _fileRepository.GetFileByIdAsync(fileId);
 
+        if (fileDal is null)
+        {
+            throw new FileNotFoundException($"File with id {fileId} not found");
+        }
+
         var fileModel = _mapper.Map<FileModel>(fileDal);
-        var fileStream = await _fileStorageService.GetFileDataAsync(fileModel);
         
+        var fileBytes = await _fileStorageService.GetFileBytesAsync(fileModel);
+
+        if (fileBytes is null)
+        {
+            throw new ObjectStorageException();
+        }
+
         var fileByIdResult = _mapper.Map<GetFileResult>(fileDal);
-        fileByIdResult.FileStream = fileStream;
+        fileByIdResult.FileBytes = fileBytes;
         
         return fileByIdResult;
     }
@@ -68,8 +81,8 @@ internal class FileManager : IFileManager
     /// <inheritdoc cref="IFileManager.UpdateFileAsync"/>
     public async Task UpdateFileAsync(UpdateFileParam updateFileParam, IFormFile file)
     {
-        var fileName = FileHelper.GetFileNameFromFileNameWithExtension(file.FileName);
-        var fileExtension = FileHelper.GetExtensionFile(file.FileName);
+        var fileName = Path.GetFileNameWithoutExtension(file.FileName);
+        var fileExtension = Path.GetExtension(file.FileName);
         
         var fileDal = new FileDal
         {
@@ -90,7 +103,12 @@ internal class FileManager : IFileManager
     public async Task DeleteFileAsync(Guid fileId)
     {
         var fileDal = await _fileRepository.GetFileByIdAsync(fileId);
-        
+
+        if (fileDal is null)
+        {
+            throw new FileNotFoundException($"File with id {fileId} not found");
+        }
+
         await _fileRepository.DeleteFileAsync(fileId);
         
         var fileModel = _mapper.Map<FileModel>(fileDal);
